requirements.txt
PySide6>=6.0.0
Pillow>=9.0.0
PyMuPDF>=1.20.0
ezdxf>=1.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0 